core: move PacketCodecFactory out
Remove PacketCodecFactory out of core and connection context builder.
Now the factory resides in app-utils.
The packet codec should be built and provided from an external source to each Connection.
