+++ lightway-core/src/packet_codec.rs
@@ -70,19 +70,3 @@ pub type WeakPacketEncoderType = Weak<dyn PacketEncoder + Send + Sync>;
 pub type PacketDecoderType = Arc<dyn PacketDecoder + Send + Sync>;
 /// Weak reference for [`PacketDecoder`]
 pub type WeakPacketDecoderType = Weak<dyn PacketDecoder + Send + Sync>;
-
-/// Factory to build [`PacketEncoderType`] and [`PacketDecoderType`]
-/// This will be used to build a new instance of [`PacketEncoderType`] and [`PacketDecoderType`] for every connection.
-pub trait PacketCodecFactory {
-    /// Build a new instance of [`PacketEncoderType`]
-    fn build_encoder(&self) -> PacketEncoderType;
-
-    /// Build a new instance of [`PacketDecoderType`]
-    fn build_decoder(&self) -> PacketDecoderType;
-
-    /// Returns the codec name for debugging purpose
-    fn get_codec_name(&self) -> String;
-}
-
-/// Type for [`PacketCodecFactory`]
-pub type PacketCodecFactoryType = Box<dyn PacketCodecFactory + Send + Sync>;
